add parameter $manage_nodejs_package
Defaults to true. This allows for using this module without managing
the nodejs package. The use case for this is where you want to use the
npm package provider in this module, but otherwise have no need for
directly installing nodejs. This patch allows for expressing that use
case in your puppet code.